VB-1358 - move nested classes out of CancelVisitTest as these tests were not working correctly.
